perf: keep Automerge dormant in the PWA
